pyproject.toml
[project]
name = "gasproject"
version = "0.1.0"
description = ""
authors = [
    {name = "Nicolas",email = "nicolascnagel@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas (>=2.3.3,<3.0.0)",
    "faker (>=38.0.0,<39.0.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pandera (>=0.26.1,<0.27.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
